pyproject.toml
[project]
name = "scriptkill-ai-backend"
version = "0.1.0"
description = "scriptkill project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn[standard]>=0.34.2",
    "motor>=3.7.1",
    "pymongo>=4.13.0",
    "python-jose[cryptography]>=3.4.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.20",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "email-validator>=2.2.0",
    "bcrypt>=4.3.0",
    "beanie>=1.29.0",
]
